TodoListItem component added. css updates for low light
