+++ public/index.html
@@ -24,12 +24,13 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link rel="stylesheet" href="../src/index.css"
     <title>React App</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <p>I'm regular old html</p>
+    
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
