+++ src/TodoList.js
@@ -1,29 +1,29 @@
 import React from "react";
+import TodoListItem from "./TodoListItem";
 
-const todoList = [
-    {
-      id: 1,
-      title: "wash face"
-    },
-    {
-      id: 2,
-      title: "make coffee"
-    },
-    {
-      id: 3,
-      title: "physical therapy"
-    }
-  ];
+function ToDoList(props) {
 
+  // const todoList = [
+  //     {
+  //       id: 1,
+  //       title: "wash face"
+  //     },
+  //     {
+  //       id: 2,
+  //       title: "make coffee"
+  //     },
+  //     {
+  //       id: 3,
+  //       title: "physical therapy"
+  //     }
+  //   ];
 
-  function ToDoList() {
     return (
         <ul>
-          {todoList.map(function (item) {
+          {props.todoList.map(function (item) {
             return (
-              <li key={item.objectID}>
-                {item.title}
-              </li>
+              <TodoListItem item={item}/>
+ 
             );
           })}
         </ul>
